requirements.txt
adal==1.2.2
alembic==1.0.11
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.4.2
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.25.3
Werkzeug==0.15.4
